reservation-service/requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pyjwt==2.9.0
httpx==0.27.2
python-dotenv==1.0.1
uvicorn==0.30.6